# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sample API description",
        "title": "MY Sample API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/customer": {
            "get": {
                "description": "Get home page details.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get home page.",
                "operationId": "get-home-page",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/customer/profile": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Get details about a customer.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get information about a customer.",
                "operationId": "customer-profile",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/customer/sign-in": {
            "post": {
                "description": "Customer sign in.",
                "produces": [
                    "application/json"
                ],
                "summary": "Customer sign in process.",
                "operationId": "customer-sign-in",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/customer/signup": {
            "post": {
                "description": "Create a new customer.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new customer.",
                "operationId": "create-customer",
                "parameters": [
                    {
                        "description": "Customer information to create",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/router_manager.CustomerInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/router_manager.CustomerResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "router_manager.CustomerInput": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "router_manager.CustomerResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}